### 2020-04-30 #34 @sje30
10 year challenge: Eglen and Wong (2008)
**Original article:** https://arxiv.org/abs/1208.0986 (postprint of 2008 paper)

**Scientific domain:**  Computational Neuroscience
**Programming language:**  R
**Suggested editor:**  Timothée Poisot or Karthik Ram or Etienne Roesch
