### 2020-04-30 #34 @sje30
Apologies in advance for grafting your pdf onto the front of my pdf -- I ran out of time to investigate if your article style could be used with Rmarkdown.  So creating article2.pdf was my solution.
