### 2020-04-30 sje30/rescience-hor#1 @sje30
Submission
Have made initial submission to Rescience

but am not happy at having to pdfcat my article on the back of the article.pdf from the metadata.  Is there are a better way of doing this?
https://github.com/ReScience/submissions/issues/34
