### 2020-05-04 #34 @rougier
Thanks for your submission, we'll assign an editor soon. For the published version (if accepted), that would be great if you can use the template.

I'll edit your submission.
