### 2020-05-04 #34 @sje30
> Thanks for your submission, we'll assign an editor soon. For the
> published version (if accepted), that would be great if you can use
> the template.

yes, of course, will be able to swtich to your template.  I think I can
just export the markdown to .tex and then fix up from there.
