### 2020-06-24 #34 @rgutzen
Sorry, I'm afraid I have no experience with R.
