### 2020-06-29 #34 @iampritishpatil
Sorry I missed the tag last time. I don't think I'd manage well as I haven't used R in several years, and seems there are a non trivial amount of code and dependencies in R. I'm sorry.
