### 2020-06-30 #34 @rcaze
Sorry for not answering earlier, I just started to re-enabled notification on github. I'm afraid my knowledge of R is also insufficient.
