### 2020-06-30 #34 @benoit-girard
I haven't used R in the last 17 years, I fear I won't be able to examine the code...
