### 2020-06-30 #34 @rougier
@thmosqueiro @ChrisRackauckas Could you review this submission (https://github.com/ReScience/submissions/issues/34) ?
