### 2020-07-02 #34 @rougier
@ChrisRackauckas @benoit-girard Thanks for the answer.
@eroesch stephanmg degoldschmidt Could you review this article (R + Computational Neuroscience) ?
