### 2020-09-08 #34 @rougier
Sorry for the delay. @eroesch @stephanmg @degoldschmidt Could you review this article (R + Computational Neuroscience) ?
