### 2021-05-28 #34 @rougier
Oh my god, I'm very late on following this submission.  @ChrisRackauckas @thmosqueiro @eroesch @stephanmg @degoldschmidt Could you review this article (R + Computational Neuroscience) ?
