### 2021-05-28 #34 @sje30
Thank you both!  I had forgotten about this.  I just checked the repo, and it looks like the binder broke sometime in the last year (rmarkdown needs to be installed, and there is a small pandoc-citeproc issue).  Can you give me the weekend to fix this, which in the interim will update it to R version 4.0 and handle the change in pandoc-citeproc which is now missing from Ubuntu (this will teach me not to pin all the versions down for mybinder!)
