### 2021-05-29 #34 @sje30
The `launch binder` button on the readme of https://github.com/sje30/rescience-hor is working again; just needed a couple of tweaks to system R images that had changed in the last year.
