### 2021-05-31 #34 @rougier
Sure, have a look at: https://rescience.github.io/edit/

Note that this submission is related to the "Ten Years Reproducibility Challenge" where authors try to reproduce (re-run) their own code. You can also browse [open reviews](https://github.com/ReScience/submissions/issues?q=is%3Aissue+label%3A%22Ten+Years+Challenge%22+is%3Aclosed) on the same issue.
