### 2021-05-31 #34 @stephanmg
@sje30 I looked at your reproduction.

I have some suggestions for (minor) revision before accepting the submission:
Figures:
- Figure 1: Original figure is black/white, I would avoid using red/green colors for the dots - will be much easier to discriminate for people with color vision deficiency.

Code:
-  `Error: Status code 504 returned` - I assume this some Binder-related issue?
- `hor_bdmin.R` - contains some commented code which seems not necessary, can this be removed?
- `maps2.R`  - same as above
- Overall many empty lines and superfluous white space, could you correct this?

Altogether great replication of figures and tables.  This looks good!

**Review criteria:**
1. Reproducibility of the replication: Yes, I can reproduce via Binder and on my computer.
2. Clarity and completeness of the accompanying article: Great, everything ran OOTB for me.
3. Clarity of the code: Okay.
4. Clarity and completeness of the accompanying article: Okay.
