### 2021-05-31 #34 @sje30
Thanks for doing this so promptly!

> - Figure 1: Original figure is black/white, I would avoid using
> red/green colors for the dots - will be much easier to discriminate
> for people with color vision deficiency.

The 'red' is actually 'orangered' for the exact reason you mention.  I
can change it to blue though if you prefer?

> Code:
> -  `Error: Status code 504 returned` - I assume this some
> Binder-related issue?

Hmm... the binder just worked for me now.  Could you please try again?

> - `hor_bdmin.R` - contains some commented code which seems not
> necessary, can this be removed?

> - `maps2.R`  - same as above
I can certainly tidy it up!  great idea.

> - Overall many empty lines and superfluous whitespace, could you
> correct this?

Is that in the R files?

> Altogether great replication of figures and tables. This looks okay.

Thank you.
