### 2021-05-31 #34 @stephanmg
> Thanks for doing this so promptly!
> - Figure 1: Original figure is black/white, I would avoid using red/green colors for the dots - will be much easier to discriminate for people with color vision deficiency.
> The 'red' is actually 'orangered' for the exact reason you mention. I can change it to blue though if you prefer?

Well, call me pedantic, I would just prefer black and white as in the original figure.
> Code: - `Error: Status code 504 returned` - I assume this some Binder-related issue?
> Hmm... the binder just worked for me now. Could you please try again?

Will try again later.

> - `hor_bdmin.R` - contains some commented code which seems not necessary, can this be removed? - `maps2.R` - same as above
> I can certainly tidy it up! great idea.
> - Overall many empty lines and superfluous whitespace, could you correct this?
> Is that in the R files?
Yes, in the R files, there is a lot of empty lines, sometimes in sequence.
> Altogether great replication of figures and tables. This looks okay.
> Thank you.

Besides this, I would suggest to tidy up the repository into code and markup if possible.
