### 2021-05-31 #34 @stephanmg
By the way: I am a little confused about the following: You write that the arXiv paper is a post print, do you mean preprint? Because I see your paper being published in `Vis Neurosci.`  (and I would call this a post print) and arXiv is obviously a preprint server.
