### 2021-07-26 ba2ec6e
Revert figure 1 to black/white
Avoid any potential ambiguity with red/green colorblindness.
