+++ template-master/metadata.yaml
@@ -9,15 +9,14 @@ title: "[Re] Eglen and Wong (2008) --- Spatial constraints
   underlying the retinal mosaics of two types of horizontal cells in
   cat and macaque"
 
-
 # List of authors with name, orcid number, email and affiliation
 # Affiliation "*" means contact author (required even for single-authored papers)
 authors:
   - name: Stephen J. Eglen
     orcid: 0000-0001-8607-8025
     email: sje30@cam.ac.uk
-    affiliations: 1,*      # * is for contact author
-
+    affiliations: 1
+    
 # List of affiliations with code (corresponding to author affiliations), name
 # and address. You can also use these affiliations to add text such as "Equal
 # contributions" as name (with no address).
@@ -25,6 +24,7 @@ affiliations:
   - code:    1
     name:    Department of Applied Mathematics and Theoretical Physics
     address: Cambridge, UK
+    
 # List of keywords (adding the programming language might be a good idea)
 keywords: rescience c, R
 
@@ -34,13 +34,6 @@ keywords: rescience c, R
 #   see https://guides.github.com/activities/citable-code/
 code:
   - url: https://github.com/sje30/rescience-hor
-  - doi: 
-  - swh: 
-
-# Data URL and DOI (optional if no data)
-data:
-  - url:
-  - doi:
 
 # Information about the original article that has been replicated
 replication:
@@ -68,7 +61,7 @@ type: Replication
 
 # Scientific domain of the article (e.g. Computational Neuroscience)
 #  (one domain only & try to be not overly specific)
-domain: Computational Neuroscience
+domain:  Computational Neuroscience
 
 # Coding language (main one only if several)
 language: R
