+++ template-master/bibliography.bib
@@ -1,88 +1,127 @@
-@article{Rougier:2018,
-  doi = {10.1038/d41586-018-04628-w},
-  year = {2018},
-  month = apr,
-  publisher = {Springer Nature},
-  volume = {556},
-  number = {7701},
-  pages = {309--309},
-  author = {Rougier, Nicolas P. and Hinsen, Konrad},
-  title = {Code reviewing puts extra demands on referees },
-  journal = {Nature},
+@ARTICLE{Eglen2008-ca,
+  title    = "Spatial constraints underlying the retinal mosaics of two types
+              of horizontal cells in cat and macaque",
+  author   = "Eglen, Stephen J and Wong, James C T",
+  abstract = "Most types of retinal neurons are spatially positioned in
+              non-random patterns, termed retinal mosaics. Several
+              developmental mechanisms are thought to be important in the
+              formation of these mosaics. Most evidence to date suggests that
+              homotypic constraints within a type of neuron are dominant, and
+              that heterotypic interactions between different types of neuron
+              are rare. In an analysis of macaque H1 and H2 horizontal cell
+              mosaics, W{\"a}ssle et al. (2000) suggested that the high
+              regularity index of the combined H1 and H2 mosaic might be caused
+              by heterotypic interactions during development. Here we use
+              computer modeling to suggest that the high regularity index of
+              the combined H1 and H2 mosaic is a by-product of the basic
+              constraint that two neurons cannot occupy the same space. The
+              spatial arrangement of type A and type B horizontal cells in cat
+              retina also follow this same principle.",
+  journal  = "Vis. Neurosci.",
+  volume   =  25,
+  number   =  2,
+  pages    = "209--214",
+  month    =  mar,
+  year     =  2008,
+  issn     = "0952-5238, 1469-8714",
+  pmid     = "18334045",
+  doi      = "10.1017/S0952523808080176"
 }
 
-@article{Science:2018,
-  doi = {10.1126/science.359.6377.725},
-  author = {Hutson, Matthew},
-  url = {http://science.sciencemag.org/content/359/6377/725},
-  year = {2018},
-  month = feb,
-  volume = {359},
-  number = {6377},
-  title = {Artificial intelligence faces a replication crisis},
-  journal = {Science},
+
+@ARTICLE{Wassle1978-cd,
+  title    = "Topography of horizontal cells in the retina of the domestic cat",
+  author   = "W{\"a}ssle, H and Peichl, L and Boycott, B B",
+  abstract = "Neurofibrillar methods stain a class of horizontal cells in the
+              cat retina which are shown to be identical with the A-type
+              horizontal cell of Golgi-staining. Thus all of the A-type cells
+              of a single retina can be observed. On this basis the changes in
+              density and dendritic field size of A-type horizontal cells with
+              respect to retinal eccentricity were measured. The decrease in
+              density from centre to periphery is balanced by a corresponding
+              increase in size of the dendritic field. Consequently each
+              retinal point--independent of retinal position--is covered by the
+              dendritic fields of three of four A-type horizontal cells. The
+              nuclei and nucleoli of B-type horizontal cells could also be
+              recognized in neurofibrillar-stained material and thus their
+              distribution was determined. The density ratio B-type: A-type is
+              2.8 +/- 0.4 and does not vary much from the centre to the
+              periphery of the retina. Each retinal point is also covered by
+              four B-type horizontal cells. Thus a single cone can contact a
[…]
