### 2021-07-26 #34 @sje30
I have reverted figure 1 to b/w, and removed excess comments.

I'm now getting problems however getting article.tex to compile.  I think it may be my lack of python skills, but when I run make in the template_master folder, I get the following error:

```
make
./yaml-to-latex.py -i metadata.yaml -o metadata.tex
Traceback (most recent call last):
  File "/home/stephen/proj/2020/rescience-hor/template-master/./yaml-to-latex.py", line 93, in <module>
    content = generate_latex_metadata(filename_in, article)
  File "/home/stephen/proj/2020/rescience-hor/template-master/./yaml-to-latex.py", line 10, in generate_latex_metadata
    "% DO NOT EDIT - automatically generated from {filename}\n\n"
AttributeError: 'str' object has no attribute 'name'
make: *** [Makefile:18: metadata.tex] Error 1
```

I did also have to fix article.py to specify which loader to use.
