### 2021-07-26 #34 @sje30
```
Python 3.9.6 (default, Jun 30 2021, 10:22:16)
[GCC 11.1.0] on linux
>>> import yaml
>>> print (yaml.__version__)
5.4.1
```
