+++ template-master/content.tex
@@ -1,1 +1,257 @@
-Please see attached submission on following pages.
+\subsection{Summary}\label{summary}}
+
+This paper reports the successful reproduction of the results in a paper
+I co-authored in 2008 (Eglen and Wong 2008). A postprint of this paper
+is available at \url{https://arxiv.org/abs/1208.0986}. (Open access was
+not mandated by the funders of this research at the time of publication,
+2008.) All the code has been uploaded to
+\url{https://github.com/sje30/rescience-hor} along with the Rmarkdown
+sources for this report.
+
+\hypertarget{historical-context}{%
+\section{Historical context}\label{historical-context}}
+
+\hypertarget{scientific-question}{%
+\subsection{Scientific question}\label{scientific-question}}
+
+To investigate a problem in neurobiology concerning the relative spatial
+positioning of two types of neuron in the retina. We used three example
+datasets (pretty much all that was available at the time) to investigate
+this problem using spatial statistics and simulations of spatial
+patterning. These datasets were not available in digital format, but the
+images (Wässle, Peichl, and Boycott 1978; Wässle et al. 2000) were
+scanned in and digitised (each cell represented as an x,y location) by
+me as part of the work for the original paper. These datasets were not
+publicly shared before.
+
+\hypertarget{computational-context}{%
+\subsection{Computational context}\label{computational-context}}
+
+Most of the code required for the 2008 paper had been developed by
+myself since about 2000 working on other related problems. The major
+innovation here was in the biological problem.
+
+One of the key reasons I switched to R as a programming environment in
+the late 1990s was because of the strong support it had for spatial
+statistics. In particular the SPLANCS package (Rowlingson and Diggle
+2017) in R has been a key package in most of my research.
+
+\hypertarget{original-source-code}{%
+\subsection{Original source code}\label{original-source-code}}
+
+the original source code was not published or publicly archived before.
+The source code was kept just on my hard drive. I keep good backups
+though, and in this case the projects were still in my home directory
+(copied from machine to machine over the years). No license was written
+systematically to cover the code I wrote, although the R packages are
+usually released under GPL or MIT.
+
+As a side-effect of producing this article, the following R packages of
+mine were made available:
+
+\begin{enumerate}
+\def\labelenumi{\arabic{enumi}.}
+\tightlist
+\item
+  sjedmin: dmin model for spatial point patterns.
+  \url{https://github.com/sje30/sjedmin}
+\item
+  sjedist: unified calculation of spatial statistics.
+  \url{https://github.com/sje30/sjedist}
+\item
+  sjevor: Wrapper around public domain code for Voronoi tessellation
+  \url{https://github.com/sje30/sjevor}
+\end{enumerate}
+
+The Voronoi tessellation was public domain C code provided by Steve
+Fortune (Fortune 1987). (This code was not required for the current
+project, but was implicitly required for the parent package, sjedist.)
+
+\hypertarget{retrieval-of-the-software}{%
+\subsection{Retrieval of the software}\label{retrieval-of-the-software}}
+
+This was relatively straightforward, as the software was all on my hard
+drive. One complication was that until a few years ago, I had been using
+RCS as my primary version control software. As I wished to release the
+packages onto github, I converted the history from RCS into Git using
+the tool \texttt{https://github.com/Oblomov/rcs-fast-export}.
+
[…]
