### 2021-07-27 #34 @rougier
Can you check if you have the latest version of the article template (with `yaml-to-latex.py`)?
