+++ template/metadata.tex
@@ -1,0 +1,42 @@
+% DO NOT EDIT - automatically generated from metadata.yaml
+
+\def \codeURL{https://github.com/sje30/rescience-hor}
+\def \codeDOI{}
+\def \codeSWH{}
+\def \dataURL{}
+\def \dataDOI{}
+\def \editorNAME{}
+\def \editorORCID{}
+\def \reviewerINAME{}
+\def \reviewerIORCID{}
+\def \reviewerIINAME{}
+\def \reviewerIIORCID{}
+\def \dateRECEIVED{01 November 2018}
+\def \dateACCEPTED{}
+\def \datePUBLISHED{}
+\def \articleTITLE{[Re] Eglen and Wong (2008) --- Spatial constraints underlying the retinal mosaics of two types of horizontal cells in cat and macaque}
+\def \articleTYPE{Replication}
+\def \articleDOMAIN{Computational Neuroscience}
+\def \articleBIBLIOGRAPHY{bibliography.bib}
+\def \articleYEAR{2021}
+\def \reviewURL{}
+\def \articleABSTRACT{Three years have passed since ReScience published its first article and since September 2015, things have been going steadily. We're still alive, independent and without a budget. In the meantime, we have published around 20 articles (mostly in computational neuroscience \& computational ecology) and the initial has grown from around 10 to roughly 100 members (editors and reviewers), we have advertised ReScience at several conferences worldwide, gave some interviews and we published an article introducing ReScience in PeerJ. Based on our experience at managing the journal during these three years, we think the time is ripe for proposing some changes.}
+\def \replicationCITE{}
+\def \replicationBIB{}
+\def \replicationURL{}
+\def \replicationDOI{}
+\def \contactNAME{Stephen J. Eglen}
+\def \contactEMAIL{sje30@cam.ac.uk}
+\def \articleKEYWORDS{retinal mosaics, spatial statistics, R}
+\def \journalNAME{ReScience C}
+\def \journalVOLUME{4}
+\def \journalISSUE{1}
+\def \articleNUMBER{}
+\def \articleDOI{}
+\def \authorsFULL{Stephen J. Eglen}
+\def \authorsABBRV{S.J. Eglen}
+\def \authorsSHORT{Eglen}
+\title{\articleTITLE}
+\date{}
+\author[1,\orcid{0000-0001-8607-8025}]{Stephen J. Eglen}
+\affil[1]{Department of Applied Mathematics and Theoretical Physics, Cambridge, UK}

[…]
