### 2021-07-31 f7675fa
remove template bib entries
