+++ template/bibliography.bib
@@ -128,91 +128,3 @@
 
 
 
-@article{Rougier:2018,
-  doi = {10.1038/d41586-018-04628-w},
-  year = {2018},
-  month = apr,
-  publisher = {Springer Nature},
-  volume = {556},
-  number = {7701},
-  pages = {309--309},
-  author = {Rougier, Nicolas P. and Hinsen, Konrad},
-  title = {Code reviewing puts extra demands on referees },
-  journal = {Nature},
-}
-
-@article{Science:2018,
-  doi = {10.1126/science.359.6377.725},
-  author = {Hutson, Matthew},
-  url = {http://science.sciencemag.org/content/359/6377/725},
-  year = {2018},
-  month = feb,
-  volume = {359},
-  number = {6377},
-  title = {Artificial intelligence faces a replication crisis},
-  journal = {Science},
-}
-
-@Misc{Roboto:2011,
-  author =       {Christian Robertson},
-  title =        {The Roboto family of fonts (Google)},
-  url =          {https://github.com/google/roboto},
-  year =         2011,
-  note =         {Apache License, verison 2.0},
-}
-
-@Misc{SourceSerifPro:2014,
-  author =       {Frank Grießhammer},
-  title =        {Source Serif Pro (Adobe Systems)},
-  url =          {https://github.com/adobe-fonts/source-serif-pro},
-  year =         2014,
-  note =         {SIL Open Font License, version 1.1},
-}
-
-@Misc{SourceCodePro:2012,
-  author =       {Paul D. Hunt},
-  title =        {Source Code Pro (Adobe Systems)},
-  url =          {https://github.com/adobe-fonts/source-code-pro},
-  year =         2012,
-  note =         {SIL Open Font License, version 1.1},
-}
-
-@article{Topalidou:2015,
-  author =       {Topalidou, Meropi and Rougier, Nicolas P.},
-  title =        {{[Re] Interaction between cognitive and motor cortico-basal
-                  ganglia loops during decision making: a computational study}},
-  journal =      {ReScience},
-  year =         2015,
-  volume =       1,
-  number =       1,
-  doi =          {10.5281/zenodo.27944},
-}
-
-@article{Rougier:2017,
-  doi =          {10.7717/peerj-cs.142},
-  author =       {Nicolas P. Rougier and Konrad Hinsen and Frédéric Alexandre
-                  and Thomas Arildsen and Lorena Barba and Fabien
-                  C. Y. Benureau and C. Titus Brown and Pierre de Buyl and Ozan
-                  Caglayan and Andrew P. Davison and Marc André Delsuc and
-                  Georgios Detorakis and Alexandra K. Diem and Damien Drix and
-                  Pierre Enel and Benoît Girard and Olivia Guest and Matt
-                  G. Hall and Rafael Neto Henriques and Xavier Hinaut and Kamil
-                  S Jaron and Mehdi Khamassi and Almar Klein and Tiina Manninen
-                  and Pietro Marchesi and Dan McGlinn and Christoph Metzner and
-                  Owen L. Petchey and Hans Ekkehard Plesser and Timothée Poisot
-                  and Karthik Ram and Yoav Ram and Etienne Roesch and Cyrille
-                  Rossant and Vahid Rostami and Aaron Shifman and Joseph
-                  Stachelek and Marcel Stimberg and Frank Stollmeier and
-                  Federico Vaggi and Guillaume Viejo and Julien Vitay and Anya
-                  Vostinar and Roman Yurchak and Tiziano Zito},
-  title =        {{Sustainable computational science: the ReScience
-                  initiative}},
-  journal =      {{PeerJ} Computer Science},
-  month =        12,
-  volume =       3,
[…]
