### 2021-07-31 150627b
fix up table 1 (table 2 from orig paper)
