+++ template/content.tex
@@ -37,12 +37,11 @@ package\supercite{splancs} in R has been a key package in most of my research.
 
 \subsection{Original source code}
 
-the original source code was not published or publicly archived before.
-The source code was kept just on my hard drive. I keep good backups
-though, and in this case the projects were still in my home directory
-(copied from machine to machine over the years). No license was written
-systematically to cover the code I wrote, although the R packages are
-usually released under GPL or MIT.
+The original source code was not published or publicly archived
+before. I keep good backups though, and in this case the projects were
+still in my home directory (copied from machine to machine over the
+years). No license was written systematically to cover the code I
+wrote, although the R packages are usually released under GPL or MIT.
 
 As a side-effect of producing this article, the following R packages of
 mine were made available:
