+++ README.md
@@ -17,10 +17,5 @@ the file `rescience-hor.pdf`.
 
 ## Note for the Rescience editors
 
-I have used your template to create a cover page for my submission,
-following the yaml file.  I did not yet attempt how to generate a pdf
-from Rmd using your templates.  My
-[article](template-master/article2.pdf) with
-[metadata](template-master/metadata.yaml) are in the template-master folder.
-
-
+The [article](template/article.pdf) and 
+[metadata](template/metadata.yaml) are provided in the template folder.
