### 2021-07-31 #34 @sje30
thanks.  I started over with a fresh template/ and that seemed to do the job nicely.  I have now pushed the new version to github, and converted from the Rmarkdown to use the template.  See: https://github.com/sje30/rescience-hor/blob/master/template/article.pdf
