+++ template/content.tex
@@ -32,8 +32,8 @@ innovation here was in the biological problem.
 
 One of the key reasons I switched to R as a programming environment in
 the late 1990s was because of the strong support it had for spatial
-statistics. In particular the SPLANCS
-package\supercite{splancs} in R has been a key package in most of my research.
+statistics. In particular the SPLANCS package\supercite{splancs} in R
+has been a key package in most of my research.
 
 \subsection{Original source code}
 
