### 2021-08-19 #34 @rougier
sorry for the delay. Articlel looks good to me and from the review by @stephanmg I think we can accept the paper.

I'll try to edit it today starting from your new version. Stay tuned.
