### 2021-08-19 #34 @rougier
I would need a software heritage ID. Can you save your repo at https://www.softwareheritage.org/save-and-reference-research-software/ and give me back the corresponding swid?
