### 2021-08-21 #34 @khinsen
Software heritage seems to have some technical issues at the moment, I didn't manage to save a GitHub repo yesterday. BTW, you don't need an account there, what you did should have been sufficient.
