### 2021-08-21 #34 @sje30
Thanks Konrad.  Seems to be a bit healthier this morning -- I can see this now:

https://archive.softwareheritage.org/browse/origin/directory/?origin_url=https://github.com/sje30/rescience-hor
