### 2021-08-21 #34 @sje30
@rougier : is this what you need?
swh:1:dir:6088b7312865b39a4a5f5af78556bf6da5fa6409;origin=https://github.com/sje30/rescience-hor;visit=swh:1:snp:a2df6c1cf91ec9a80d08026d7728a503eb64c7a6;anchor=swh:1:rev:0882985f8a0e3ea32f8898175460bacc8060b552
