### 2021-08-31 #34 @rougier
The sandboxed version is online at https://sandbox.zenodo.org/record/908095, can you check if everything's ok? Then I'll proceed with the actual publication.
