### 2021-08-31 #34 @rougier
It's online at https://zenodo.org/record/5347786 ! Congratulations. It should appear soon on rescience website (a few minutes)
